pyproject.toml
[project]
name = "docman"
version = "0.5.0"
description = "Document processing pipeline — Heddle framework test project"
requires-python = ">=3.11"
dependencies = [
    "heddle-ai[duckdb]",             # Heddle framework + DuckDB contrib
    "markitdown>=0.1.0",           # Fast document-to-Markdown conversion
    "docling>=2.0.0",              # Deep document extraction (PDF OCR, tables)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docman"]

[tool.uv.sources]
heddle-ai = { path = "../heddle", editable = true }

[tool.coverage.run]
source = ["docman"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 60
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "F", "E", "W", "I", "N", "UP", "B", "A", "C4", "SIM",
    "TC", "RUF", "D", "ANN", "PT", "RET", "ARG", "PL",
    "PERF", "LOG", "T20",
]
ignore = [
    "D100", "D104", "D203", "D212", "D105", "D107",
    "ANN401", "ARG002", "PLR0913", "PLR2004",
    "B008", "PLC0415",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "ANN", "ARG", "T20", "PLR", "SIM105", "PT011", "B007", "RUF059", "RUF012", "PERF102", "RET504", "N806"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["docman"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.rumdl]
exclude = [".venv", "node_modules", "_build"]
disable = [
    "MD013",  # line-length — docs use natural line breaks, not hard-wrapped at 80
    "MD024",  # duplicate-heading — repeated "Prerequisites", "Troubleshooting" across sections
    "MD034",  # no-bare-urls — bare URLs in config examples and quick references
    "MD036",  # no-emphasis-as-heading — bold text as pseudo-headings in some docs
    "MD041",  # first-line-heading — some reference docs don't start with H1
    "MD042",  # no-empty-links — status badges use empty link targets
]

[tool.rumdl.MD004]
style = "dash"

[tool.rumdl.MD029]
style = "ordered"
